+++ LeagueVineGameViewController.m
@@ -116,7 +116,7 @@
     if (indexPath.section == 0) {
         cell.textLabel.text = item ? [item listDescription] : @"Any tournament";
     } else {
-        cell.textLabel.text = item ? [item shortDescription] : [NSString stringWithFormat: @"%@ not selected", [self sectionName:indexPath.section]];
+        cell.textLabel.text = item ? [item shortDescription] : [NSString stringWithFormat: @"%@ not selected", [self sectionName:(int)indexPath.section]];
     }
     cell.textLabel.textColor = item ? [UIColor blackColor] : [UIColor grayColor];
     
@@ -132,7 +132,7 @@
     headerLabel.textColor = uirgb(134,134,134);
     headerLabel.shadowColor = [UIColor whiteColor];
     headerLabel.shadowOffset = CGSizeMake(0, 1);
-    headerLabel.text = [self sectionName:section];
+    headerLabel.text = [self sectionName:(int)section];
     [headerView addSubview:headerLabel];
     return headerView;
 }
